refactor(#181): use migrations library for TextTranslator
